responding to commenters' requests
